Major updates
I basically finished this as a first draft. I'll have to update
the README since I didn't do a great job of committing history as
I went.  Anyway, so far, the RandomForestClassifier works best,
but the score I got was a 77%.  Not terrible but not great either.

I'll save experimenting for another day.